test: Move LRP tests to a separate suite
[ upstream commit 412e299df74f8a5e1d47e318db6ee6929893ad15 ]

To increase sanity of K8sServicesTest size.
